+++ cucumber-spring/src/test/java/io/cucumber/spring/cucumbercontextconfigannotation/AnnotatedInterface.java
@@ -1,0 +1,7 @@
+package io.cucumber.spring.cucumbercontextconfigannotation;
+
+import io.cucumber.spring.CucumberContextConfiguration;
+
+@CucumberContextConfiguration
+public interface AnnotatedInterface {
+}